refactor(theme): move LiquidData to the data layer
