+++ app/src/main/java/com/osfans/trime/data/theme/LiquidData.kt
@@ -3,9 +3,8 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package com.osfans.trime.ime.symbol
+package com.osfans.trime.data.theme
 
-import com.osfans.trime.data.theme.Theme
 import com.osfans.trime.data.theme.model.LiquidKeyboard
 
 object LiquidData {
