+++ app/src/main/java/com/osfans/trime/ime/symbol/LiquidWindow.kt
@@ -13,6 +13,7 @@ import com.google.android.flexbox.FlexboxLayoutManager
 import com.osfans.trime.daemon.RimeSession
 import com.osfans.trime.daemon.launchOnReady
 import com.osfans.trime.data.SymbolHistory
+import com.osfans.trime.data.theme.LiquidData
 import com.osfans.trime.data.theme.ThemeScope
 import com.osfans.trime.data.theme.model.LiquidKeyboard
 import com.osfans.trime.ime.core.TrimeInputMethodService
